feat(tn_cache): implement circuit breaker and resolution enhancements
- Introduced a new `circuit.go` file to implement circuit breaker functionality for managing stream data refreshes, enhancing resilience against failures.
- Added methods for obtaining and refreshing stream data with circuit breaker support, improving error handling during data refresh operations.
- Created a `resolution.go` file to manage the resolution of stream specifications, including handling of wildcard and child streams, ensuring dynamic updates to cached streams.
- Enhanced the `CacheScheduler` with improved methods for resolving directives and managing stream configurations, streamlining the caching process.
- Updated tests to cover new circuit breaker and resolution functionalities, ensuring robustness and reliability.

These changes enhance the overall performance and reliability of the cache management system.
